+++ app/models/folio/user.rb
@@ -7,8 +7,6 @@ class Folio::User < Folio::ApplicationRecord
   include Folio::HasNewsletterSubscriptions
   include Folio::HasSiteRoles
 
-  has_sanitized_fields :email, :first_name, :last_name, :company_name, :nickname
-
   # used to validate before inviting from console in /console/users/new
   attribute :skip_password_validation, :boolean, default: false